Add mirror control, fix initial strain state
Can now mirror the model (about its initial x-axis) to allow models to switch left and right. Note this breaks the sense of surfaces or handedness of volumes so is really for demonstration only.
Initial state on loading was erroneously used for strains, now changed to state after alignment/scaling.